api/requirements.txt
Flask==2.2.3
Flask-Session==0.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.4
setuptools==67.6.0

